pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "2023-autumn-AB-python-HW-13"
version = "0.1.0"
description = "HW13"
authors = ["Artyom Romanov <artem.romanov.03@bk.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.25.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.flake8]
max-line-length = 120
ban-todos = true

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = "--strict-markers -W error"
junit_family = "xunit2"
testpaths = ["tests"]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 100
omit = [
#    "server/asgi.py",
#    "server/app.py",
]